refactor: extract email login into separate package
